requirements.txt
selenium==4.21.0
azure-storage-blob==12.13.0
python-dotenv==1.0.1
pillow==10.2.0
pandas==1.3.5
openai==1.20.0